Moved some code to separate files
